Refactor mapper configuration
